pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "motor>=3.7.0",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "python-multipart>=0.0.20",
    "redis>=6.0.0",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
